Documentation Update
Today documentation was updated and the GitHub errors were fixed and the Gantt chart was completed.  I have started to clean up all of the project development file and made it easy to read.  No updates impacted the program as they are simply documentation changes. Some design parts are incomplete and need to be fixed. I need to also finish off the structure chart and flowchart.