Update to documentation and code
Firstly, we have an big update to the documentation. I have added most of the design phase, some things may need a slight touch up and will be done tomorrow afternoon to stay on schedule. Secondly we have the beginning of the development phase. I have created the API fetching functions and am receiving the data from each as expected. I have also begun the Flask GUI and have started to think about the processes needed.  By beginning the development phase, I understood that I needed to focus my program in order to match time constraints and keep a level of quality. While this means I did have to take away some of the original idea it's allowing me to do a better job with the core ideas. I didn't have many problems as I have worked with flask and requests before, however I did face some query problems which led to me re-reading the documentation to solve a bug in the program. The documentation changes didn't have any effect on the code. Next I'm planning to fix some tiny details in design + add the saved image of my Flowchart and continue with the developmental phase of the project by building out the GUI further and preparing for integration.