Small GUI update
Today, I worked on the GUI of my project. I updated the html templates and did testing on how I'm going to integrate the software together. The changes were successful however I have to resolve a bug with flask pathing. They didn't quite work as expected so had to problem solve my html to get flask variables to correctly function.  Next I'm planning to fix some tiny details in design + add the saved image of my Flowchart and continue with the developmental phase of the project by building out the GUI further and preparing for integration.