HTML+App update
Worked on receiving and using parameters. All of the HTML files are updated, with some new additions such as the help file which is to be completed in the maintenance phase. I have added in path functionality that takes you to the different options. For logistical reasons I have removed the NeoWs overview as it's a ridiculously large API response that is impractical to display and furthermore NeoWs feed has the same functionality just it allows you to select the start and end date for the data. I did lots of testing in this update by using print statements, using the flask debug terminal and simply running the program. I have faced a lot of bugs in this part of the program and I fixed them by incorporating the  debugging strategies from above and some help from stack overflow. Next I need to pass the parameters from param select back to the program and complete some of the explanation comments before integrating the API into the program and creating the code that allows the graphs to work.