Major Update
I added API tables and other API values being passed to flask and sorted the API data. I also updated all of the error handling. The changes didn't work as expected and took a lot of function and system testing to debug. Next I have to fix the graphing and tables and maybe neaten them with some pretty css