Update to css and organisational changes,
The program received some pretty html updates and also received a history update in the program. I also removed some of the test cases in my program. None of the program is impacted as it was mostly cosmetic changes. Next I will continue with the CSS of the program before filling out the documentation