Bug fixing and documentational changes.
I added styling changes to the CSS, fixed the matplotlib bugs and updated most of the documentation. The program had a matplotlib bug that was fixed by changing the plotting method.  The other parts of the program changed were documentation and aesthetic changes and didn't have any problems. Next I need to peer evaluate, add in my updated flowcharts and other design charts.