Final Commit
I realised that the NeoWs graph had an issue with plotting 7 days in the final testing and had to change the program's tick rotation and alignment. I found this problem with unit testing and luckily knew a solution.  I then updated documentation and the video to be accurate. There is no next as I should be done! Will do a complete check again.